+++ src/aks-preview/azext_aks_preview/commands.py
@@ -245,6 +245,14 @@ def load_command_table(self, _):
             "delete-machines", "aks_agentpool_delete_machines", supports_no_wait=True
         )
 
+    # AKS nodepool manual-scale command
+    with self.command_group(
+        "aks nodepool manual-scale", managed_clusters_sdk, client_factory=cf_agent_pools
+    ) as g:
+        g.custom_command("add", "aks_agentpool_manual_scale_add", supports_no_wait=True)
+        g.custom_command("update", "aks_agentpool_manual_scale_update", supports_no_wait=True)
+        g.custom_command("delete", "aks_agentpool_manual_scale_delete", supports_no_wait=True)
+
     with self.command_group(
         "aks machine", machines_sdk, client_factory=cf_machines
     ) as g: