+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,17 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* Update --enable-advanced-network-observability description to note additional costs and add missing flag to create command.
+* Add etag support (--if-match, --if-none-match) to some aks commands for optimistic concurrency control.
+
+4.0.0b5
+++++++++
+* Add warnings to `az aks mesh` commands for out of support asm revision in use.
+
+4.0.0b4
+++++++++
+* Add `--vm-sizes` to `az aks create` and `az aks nodepool add`.
+* Add `az aks nodepool manual-scale add/update/delete` commands.
 
 4.0.0b3
 +++++++